src/ingestion/requirements.txt
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
pandas>=2.1.0
pyarrow>=14.0.0
tqdm>=4.66.0
